+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 As this project is pre 1.0, breaking changes may happen for minor version bumps. A breaking change will get clearly notified in this log.
 
 ## Pending
+### Update
+* Update `org.stellar.sdk.responses`, add missing fields. ([#570](https://github.com/stellar/java-stellar-sdk/pull/570))
 
 ## 0.43.0
 ### Update